+++ lib/networking/default_response.dart
@@ -6,7 +6,10 @@ class DefaultResponse<T> {
 
   late final T data;
 
-  DefaultResponse.fromJson(json, decoders, {required Type type}) {
+  DefaultResponse.fromJson(json, Map<Type, dynamic> decoders,
+      {required Type type}) {
+    assert(decoders.containsKey(type),
+        'Your config/decoders.dart file does not contain a decoder for the following class: ${type.toString()} in modelDecoders');
     data = decoders[type]!(json) as T;
   }
 }